+++ elastiknn-testing/src/test/scala/com/klibisz/elastiknn/query/MatchHashesAndScoreQuerySuite.scala
@@ -13,7 +13,7 @@ import scala.collection.mutable.ArrayBuffer
 
 class MatchHashesAndScoreQuerySuite extends FunSuite with Matchers with LuceneSupport {
 
-  val ft = new VectorMapper.FieldType("elastiknn_dense_float_vector")
+  val ft: VectorMapper.FieldType = new VectorMapper.FieldType("elastiknn_dense_float_vector")
 
   test("empty harness") {
     indexAndSearch() { (_: IndexWriter) =>
@@ -91,7 +91,7 @@ class MatchHashesAndScoreQuerySuite extends FunSuite with Matchers with LuceneSu
     }
   }
 
-  test("documents with 0 hashes are not candidates") {
+  test("documents with 0 matches are not candidates") {
     indexAndSearch() { w =>
       for (_ <- 0 until 10) {
         val d = new Document()